Renamed containing folder for default connection info. Also added serverURL to the file.
