+++ src/DefaultValues/ConfigData.jsx
@@ -1,39 +1,33 @@
-const ConfigData = [
+const ConfigData = {topics: [
   {
     name: "Camera",
     path: "/mavs_ros/image",
-    cName: "topic",
   },
   {
     name: "Global Path",
     path: "/nature/global_path",
-    cName: "topic",
   },
   {
     name: "Occupancy Grid",
     path: "/occupancy_grid_image",
-    cName: "topic",
   },
   {
     name: "Odometry",
     path: "/nature/odometry",
-    cName: "topic",
   },
   {
     name: "IMU",
     path: "/mavs_ros/imu",
-    cName: "topic",
   },
   {
     name: "Point Cloud",
     path: "/nature/points",
-    cName: "topic",
   },
   {
     name: "Waypoint",
     path: "/nature/waypoints",
-    cName: "topic",
-  }
-];
+  }], 
+  serverURL: "localhost:9090"
+};
 
 export default ConfigData;
